Fix bug in code for rethrowing exceptions in worker threads.
Current code will often miss an exception thrown by the last thread. This occurs frequently if there is only one thread (using “-thread 1”) in the thread pool.

The while loop at line 337 checks the condition “executor.getActiveCount() > 0” at the beginning of each iteration. The last active thread could have ended during the previous “Thread.sleep(1000);” and thrown an exception. But that exception will not be rethrown because “executor.getActiveCount() == 0” and we break out of the while loop.