Add a parameter to override a test and include spectra for mzML/mzXML centroided spectra
Existing code always tests the spectrum peak data for a minimum median peak distance of 50 ppm (to be considered centroid data), which always overrides the value read from an mzML or mzXML file. The parameter overrides the test result in the case that the mzML/mzXML file reports the spectrum as centroided, but the peak data failed the test.